move stuff and fix little things
