+++ deeptime/markov/msm/tram/_bindings/src/tram_module.cpp
@@ -1,7 +1,7 @@
 //  author: maaike
-#include "../include/tram.h"
-#include "../include/connected_set.h"
-#include "../include/trajectory_mapping.h"
+#include "deeptime/markov/msm/tram/tram.h"
+#include "deeptime/markov/msm/tram/connected_set.h"
+#include "deeptime/markov/msm/tram/trajectory_mapping.h"
  
 PYBIND11_MODULE(_tram_bindings, m) {
     using namespace pybind11::literals;
