+++ examples/methods/plot_tram.py
@@ -1,6 +1,7 @@
 """
 TRAM on a 1D double well
-=============================
+========================
+
 This example shows how to use the transition-based reweighting analysis method (TRAM) to estimate the free energies
 and Markov model of a simple double-well potential, sampled using umbrella sampling.
