+++ tests/src/test_tram.cpp
@@ -1,7 +1,7 @@
 #include <random>
 #include <pybind11/embed.h>
 #include <catch2/catch.hpp>
-#include "../../deeptime/markov/msm/tram/_bindings/include/tram.h"
+#include "deeptime/markov/msm/tram/tram.h"
 //
 // Created by Maaike on 01/12/2021.
 //
